dev: cambio de double a Bigdecimal, documentacion del controller y enum de comisiones

+++ src/main/java/com/atorres/nttdata/transactionmicroservice/model/dao/CreditDao.java
@@ -5,11 +5,13 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class CreditDao {
     @Id
     private String id;
-    private Double balance;
-    private Double debt;
+    private BigDecimal balance;
+    private BigDecimal debt;
 }
